model/requirements.txt
fastapi
uvicorn[standard]
lightgbm
scikit-learn
pandas
numpy
pydantic
python-dotenv
supabase
